Moved DataUtil class to lib/util.rb and rm util.rb
